:recycle: Move views service utilities to separate file
