+++ src/services/__init__.py
@@ -4,4 +4,5 @@ from .food_menu import *
 from .notifiers import *
 from .photos import *
 from .secret_messages import *
+from .views import *
 from .work import *
